requirements.txt
streamlit
openai
langchain
chromadb
pdfplumber
python-dotenv